USE uint256 type instead of uint alias

+++ contracts/zombieattack.sol
@@ -3,18 +3,18 @@ pragma solidity >=0.4.19;
 import "./zombiehelper.sol";
 
 contract ZombieAttack is ZombieHelper {
-    uint randNonce = 0;
-    uint attackVictoryProbability = 70;
+    uint256 randNonce = 0;
+    uint256 attackVictoryProbability = 70;
 
-    function randMod(uint _modulus) internal returns(uint) {
+    function randMod(uint256 _modulus) internal returns(uint256) {
         randNonce = randNonce.add(1);
-        return uint(keccak256(now, msg.sender, randNonce)) % _modulus;
+        return uint256(keccak256(now, msg.sender, randNonce)) % _modulus;
     }
 
-    function attack(uint _zombieId, uint _targetId) external onlyOwnerOf(_zombieId) {
+    function attack(uint256 _zombieId, uint256 _targetId) external onlyOwnerOf(_zombieId) {
         Zombie storage myZombie = zombies[_zombieId];
         Zombie storage enemyZombie = zombies[_targetId];
-        uint rand = randMod(100);
+        uint256 rand = randMod(100);
         if (rand <= attackVictoryProbability) {
             myZombie.winCount = myZombie.winCount.add(1);
             myZombie.level = myZombie.level.add(1);
